Organized test folders, cleaned up code, added comments
